Moved CheckPassphraseStrengthFragment to ui/activity/fragment.| #1303
