+++ FlowCrypt/src/main/java/com/flowcrypt/email/ui/activity/fragment/CheckPassphraseStrengthFragment.kt
@@ -3,7 +3,7 @@
  * Contributors: DenBond7
  */
 
-package com.flowcrypt.email.ui.activity
+package com.flowcrypt.email.ui.activity.fragment
 
 import android.content.Context
 import android.graphics.Color
